+++ packages/spacecat-shared-utils/src/index.js
@@ -28,6 +28,7 @@ export {
   isValidDate,
   isValidUrl,
   isValidUUID,
+  isValidIMSOrgId,
   toBoolean,
 } from './functions.js';