+++ packages/spacecat-shared-ahrefs-client/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [@adobe/spacecat-shared-ahrefs-client-v1.6.11](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.10...@adobe/spacecat-shared-ahrefs-client-v1.6.11) (2025-03-04)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#648](https://github.com/adobe/spacecat-shared/issues/648)) ([6bdd02c](https://github.com/adobe/spacecat-shared/commit/6bdd02ccaaaf89d4b3d463206ec1377c1ccecf4d))
+
+# [@adobe/spacecat-shared-ahrefs-client-v1.6.10](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.9...@adobe/spacecat-shared-ahrefs-client-v1.6.10) (2025-03-04)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/fetch to v4.2.0 ([#645](https://github.com/adobe/spacecat-shared/issues/645)) ([7c97dc4](https://github.com/adobe/spacecat-shared/commit/7c97dc4c4853aa183553ed90f4b0d6dc9f49b656))
+
+# [@adobe/spacecat-shared-ahrefs-client-v1.6.9](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.8...@adobe/spacecat-shared-ahrefs-client-v1.6.9) (2025-03-03)
+
+
+### Bug Fixes
+
+* **ahrefs-client:** log ahrefs costs on a single line ([#642](https://github.com/adobe/spacecat-shared/issues/642)) ([9b12042](https://github.com/adobe/spacecat-shared/commit/9b120429fde96382516f9387c719e8d6592be496))
+
+# [@adobe/spacecat-shared-ahrefs-client-v1.6.8](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.7...@adobe/spacecat-shared-ahrefs-client-v1.6.8) (2025-03-03)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#638](https://github.com/adobe/spacecat-shared/issues/638)) ([64625c2](https://github.com/adobe/spacecat-shared/commit/64625c24f1b3b7bc4a26b576155bb6bc8529ef45))
+
 # [@adobe/spacecat-shared-ahrefs-client-v1.6.7](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.6...@adobe/spacecat-shared-ahrefs-client-v1.6.7) (2025-02-16)